+++ src/Navigation.js
@@ -1,17 +1,16 @@
 import React from 'react'
-import { makeStyles } from '@material-ui/core/styles'
+import { NavLink } from 'react-router-dom'
+
 import AppBar from '@material-ui/core/AppBar'
+import IconButton from '@material-ui/core/IconButton'
+import Menu from '@material-ui/core/Menu'
+import MenuItem from '@material-ui/core/MenuItem'
+import { makeStyles } from '@material-ui/core/styles'
 import Toolbar from '@material-ui/core/Toolbar'
 import Typography from '@material-ui/core/Typography'
-import IconButton from '@material-ui/core/IconButton'
-import MenuIcon from '@material-ui/icons/Menu'
+
 import AccountCircle from '@material-ui/icons/AccountCircle'
-import MenuItem from '@material-ui/core/MenuItem'
-import Menu from '@material-ui/core/Menu'
 import MoreIcon from '@material-ui/icons/MoreVert'
-import PropTypes from 'prop-types'
-import { NavLink } from 'react-router-dom'
-import Button from '@material-ui/core/Button'
 
 const useStyles = makeStyles((theme) => ({
   root: {